--- a/skyvern/webeye/scraper/scraper.py
+++ b/skyvern/webeye/scraper/scraper.py
@@ -506,8 +506,8 @@ async def scrape_web_unsafe(
     # This also solves the issue where we can't scroll due to a popup.(e.g. geico first popup on the homepage after
     # clicking start my quote)
 
-    LOG.info("Waiting for 5 seconds before scraping the website.")
-    await asyncio.sleep(5)
+    LOG.info("Waiting for 3 seconds before scraping the website.")
+    await asyncio.sleep(3)
 
     elements, element_tree = await get_interactable_element_tree(page, scrape_exclude)
     element_tree = await cleanup_element_tree(page, url, copy.deepcopy(element_tree))