<!-- ELLIPSIS_HIDDEN -->


> [!IMPORTANT]
> Reduce execution wait times across various functions in the `skyvern` project to improve performance.
> 
>   - **Behavior**:
>     - Reduce random sleep time in `agent_step()` in `agent.py` from 1-2s to 0.5-1s.
>     - Reduce sleep time in `handle_input_text_action()` in `handler.py` from 5s to 2s.
>     - Reduce sleep time in `handle_select_option_action()` in `handler.py` from 5s to 3s.
>     - Reduce sleep time in `choose_auto_completion_dropdown()` in `handler.py` from 5s to 3s.
>     - Reduce sleep time in `scroll_down_to_load_all_options()` in `handler.py` from 5s to 3s.
>     - Reduce sleep time in `navigate_to_url()` in `browser_factory.py` from 5s to 1s.
>     - Reduce sleep time in `scrape_web_unsafe()` in `scraper.py` from 5s to 3s.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 98dd1898965eb1a6f652242e46a0ab043b6cd3b2. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>


<!-- ELLIPSIS_HIDDEN -->